cleaning and rename some options (clusterName, url ...)
