+++ src/module.ts
@@ -1,12 +1,12 @@
 import { PanelPlugin } from '@grafana/data';
 import { StatusPanel } from './components/StatusPanel';
 import { statusMigrationHandler } from 'lib/statusMigrationHandler';
-import { statusPanelOptionsBuilder, StatusPanelOptions } from 'lib/statusPanelOptionsBuilder';
-import { statusFieldOptionsBuilder, StatusFieldOptions } from 'lib/statusFieldOptionsBuilder';
-import { statusPanelThresholdBuilder } from './lib/thresholdOptionsBuilder';
+import { StatusPanelOptions } from './interfaces/statusPanelOptions';
+import { StatusFieldOptions } from './interfaces/statusFieldOptions';
+import { statusPanelOptionsBuilder } from './lib/statusPanelOptionsBuilder';
+import { statusFieldOptionsBuilder } from './lib/statusFieldOptionsBuilder';
 
 export const plugin = new PanelPlugin<StatusPanelOptions, StatusFieldOptions>(StatusPanel)
   .setMigrationHandler(statusMigrationHandler)
   .setPanelOptions(statusPanelOptionsBuilder)
-  .setPanelOptions(statusPanelThresholdBuilder)
   .useFieldConfig({ useCustomConfig: statusFieldOptionsBuilder });
